requirements.txt
# python3.12
# pip install -r requirements.txt
allure-pytest==2.13.5
pylightxl==1.61
PyMySQL==1.1.1
pytest==8.3.3
pyotp==2.9.0
httpx==0.28.0
wcwidth==0.2.13
paramiko==3.5.0
kubernetes==31.0.0
redis == 5.2.0
json_tools == 0.4.1